修改区域Id类型并验证

+++ src/main/java/com/xqc/campusshop/dao/AreaDao.java
@@ -39,5 +39,5 @@ public interface AreaDao {
 	 * @param areaIdList
 	 * @return
 	 */
-	int batchDeleteArea(List<Integer> areaIdList);
+	int batchDeleteArea(List<Long> areaIdList);
 }
